landing page buttons added
